+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- Requires `innmind/filesystem:~7.4`
+
 ## 1.1.0 - 2023-09-23
 
 ### Added